AngularJS directives are extended HTML attributes with the prefix ng-. The ng-app directive initializes an AngularJS application. The ng-init directive initializes application data. The ng-model directive binds the value of HTML controls (input, select, textarea) to application data
The ng-app directive also tells AngularJS that the <div> element is the "owner" of the AngularJS application.

New directives are created by using the .directive function.
To invoke the new directive, make an HTML element with the same tag name as the new directive.
When naming a directive, you must use a camel case name, w3TestDirective, but when invoking it, you must use - separated name, w3-test-directive:

The legal restrict values are:
* E for Element name
* A for Attribute
* C for Class
* M for Comment
By default the value is EA, meaning that both Element names and attribute names can invoke the directive.